# src/tsoa-auto/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Project": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"start_date": {
						"type": "string"
					},
					"end_date": {
						"type": "string"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/Skill"
						},
						"type": "array"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"department": {
						"$ref": "#/components/schemas/Department"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Employee": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"date_of_joining": {
						"type": "string",
						"format": "date-time"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"department": {
						"$ref": "#/components/schemas/Department"
					},
					"branch": {
						"$ref": "#/components/schemas/Branch"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"buddies": {
						"items": {
							"$ref": "#/components/schemas/Buddies"
						},
						"type": "array"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/Skill"
						},
						"type": "array"
					},
					"leaverequests": {
						"items": {
							"$ref": "#/components/schemas/LeaveRequest"
						},
						"type": "array"
					},
					"hrletters": {
						"items": {
							"$ref": "#/components/schemas/HRLetters"
						},
						"type": "array"
					},
					"clockins": {
						"items": {
							"$ref": "#/components/schemas/Clockin"
						},
						"type": "array"
					},
					"project": {
						"$ref": "#/components/schemas/Project"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"role_name": {
						"type": "string"
					},
					"role_description": {
						"type": "string"
					},
					"employee": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/Permission"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Permission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"permission_name": {
						"type": "string"
					},
					"permission_description": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/Role"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Department": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"branch": {
						"$ref": "#/components/schemas/Branch"
					},
					"projects": {
						"items": {
							"$ref": "#/components/schemas/Project"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Branch": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"departments": {
						"items": {
							"$ref": "#/components/schemas/Department"
						},
						"type": "array"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Company": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"logo_url": {
						"type": "string"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"branches": {
						"items": {
							"$ref": "#/components/schemas/Branch"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Buddies": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"buddy_group_name": {
						"type": "string"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					},
					"buddy_tasks": {
						"items": {
							"$ref": "#/components/schemas/BuddyTask"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"BuddyTask": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"buddy_task": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"buddies": {
						"$ref": "#/components/schemas/Buddies"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Skill": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"projects": {
						"items": {
							"$ref": "#/components/schemas/Project"
						},
						"type": "array"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LeaveRequest": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"reason": {
						"type": "string"
					},
					"from_date": {
						"type": "string"
					},
					"to_date": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"employee": {
						"$ref": "#/components/schemas/Employee"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"HRLetters": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"letter_subject": {
						"type": "string"
					},
					"letter_content": {
						"type": "string"
					},
					"letter_time": {
						"type": "string"
					},
					"employee": {
						"$ref": "#/components/schemas/Employee"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Clockin": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"clock_in": {
						"type": "string",
						"format": "date-time"
					},
					"clock_out": {
						"type": "string",
						"format": "date-time"
					},
					"employee": {
						"$ref": "#/components/schemas/Employee"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Task": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"project": {
						"$ref": "#/components/schemas/Project"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResCompany": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"logo_url": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResBranch": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"company": {
						"$ref": "#/components/schemas/ResCompany"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResDepartment": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResRole": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"roleName": {
						"type": "string"
					},
					"roleDescription": {
						"type": "string"
					},
					"permission": {
						"items": {
							"$ref": "#/components/schemas/ResPermission"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResPermission": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"permission_name": {
						"type": "string"
					},
					"permission_description": {
						"type": "string"
					},
					"role": {
						"items": {
							"$ref": "#/components/schemas/ResRole"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResSkill": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResEmployee": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"username": {
						"type": "string"
					},
					"salary": {
						"type": "number",
						"format": "double"
					},
					"password": {
						"type": "string"
					},
					"department": {
						"$ref": "#/components/schemas/ResDepartment"
					},
					"branch": {
						"$ref": "#/components/schemas/ResBranch"
					},
					"company": {
						"$ref": "#/components/schemas/ResCompany"
					},
					"role": {
						"$ref": "#/components/schemas/ResRole"
					},
					"skills": {
						"items": {
							"$ref": "#/components/schemas/ResSkill"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ResBuddy": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"buddy_group_name": {
						"type": "string"
					},
					"employees": {
						"items": {
							"$ref": "#/components/schemas/ResEmployee"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResBuddyTask": {
				"properties": {
					"buddies": {
						"$ref": "#/components/schemas/ResBuddy"
					},
					"buddyTask": {
						"type": "string"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReqBuddyTask": {
				"properties": {
					"buddyTask": {
						"type": "string"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "boolean"
					},
					"buddy": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReqBuddy": {
				"properties": {
					"buddy_group_name": {
						"type": "string"
					},
					"buddies": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"buddies"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Api-Token": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "WorkHive APIs",
		"version": "0.0.1",
		"description": "WorkhiveServer",
		"contact": {}
	},
	"paths": {
		"/project/{projectId}/task/{projectId}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/buddy/{buddyId}/buddyTask/jwt": {
			"get": {
				"operationId": "Jwtthing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"department": {},
										"branch": {},
										"company": {},
										"id": {}
									},
									"required": [
										"department",
										"branch",
										"company",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"BuddyTask"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": []
			}
		},
		"/buddy/{buddyId}/buddyTask": {
			"get": {
				"operationId": "GetBuddyTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ResBuddyTask"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"BuddyTask"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "buddyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "SaveBuddyTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResBuddyTask"
								}
							}
						}
					}
				},
				"tags": [
					"BuddyTask"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "buddyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqBuddyTask"
							}
						}
					}
				}
			}
		},
		"/buddy/{buddyId}/buddyTask/{buddyTaskId}": {
			"put": {
				"operationId": "UpdateBuddyTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResBuddyTask"
								}
							}
						}
					}
				},
				"tags": [
					"BuddyTask"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "buddyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "buddyTaskId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqBuddyTask"
							}
						}
					}
				}
			}
		},
		"/buddy": {
			"post": {
				"operationId": "SaveBuddies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResBuddy"
								}
							}
						}
					}
				},
				"tags": [
					"Buddies"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqBuddy"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllBuddies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ResBuddy"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Buddies"
				],
				"security": [],
				"parameters": []
			}
		},
		"/buddy/yourbuddies": {
			"get": {
				"operationId": "GetYourBuddies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResBuddy"
								}
							}
						}
					}
				},
				"tags": [
					"Buddies"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": []
			}
		},
		"/buddy/{buddyId}": {
			"put": {
				"operationId": "UpdateBuddies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResBuddy"
								}
							}
						}
					}
				},
				"tags": [
					"Buddies"
				],
				"security": [
					{
						"Api-Token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "buddyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqBuddy"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}